feat(health): dedicated /alert-settings page — shared settings panel with dialog
Extract the Health Settings dialog body into HealthSettingsPanel and render
it both in the dialog (/health?settings=alerts still works) and on a new
full /alert-settings page with ?tab= deep links. Sidebar Alert Settings
entry now points at the page.
